Typo fix in zfs replace command, bump doc date
Fix a typo where the replacement disk was the same as the new disk, even
though the explanation said otherwise. Fix this by matching the last
parameter to the actual disk being replaced.

While here, bump the document date on page 2 to April.

Thanks to @romulous75 for reporting this!

Fixes #1